Fix various platform specific issues and setup CI for windows, linux and mac. (#28)
* Try multiple systems.

* Add debug ci workflow.

* Fix.

* Fix.

* Add password.

* Compilation fixes.

* Fix.

* Try fix windows.

* Fix.

* Debug on windows.

* Fix.

* Attempt.

* Fix.

* Fix.

* Minor change to libgcc static flag.

* Downgrade C++.

* Try fix cpp standard.

* Fix.

* Fix.

* Specify release config.

* Fix.

* Fix.

* Fix.

* Try MD.

* Try.

* Fix Mac.

* Fix.

* Fix.

* Explicitly enable exceptions.

* Fix artifact upload.

* Minor fix.

+++ src/preprocess.cpp
@@ -18,7 +18,7 @@ namespace coacd
 
         for (unsigned int i = 0; i < input.points.size(); ++i)
         {
-            points.push_back({input.points[i][0] * scale, input.points[i][1] * scale, input.points[i][2] * scale});
+            points.push_back({(float)(input.points[i][0] * scale), (float)(input.points[i][1] * scale), (float)(input.points[i][2] * scale)});
         }
         for (unsigned int i = 0; i < input.triangles.size(); ++i)
         {
